requirements.txt
flask==2.0.3
werkzeug==2.0.3
azure-cognitiveservices-vision-computervision==0.9.0
pillow==9.5.0
python-dotenv==1.0.0
